Implementing validateDocument
Adding the handler for document validation.
Implementing a few basic checks for the
request, getting the schema from the store,
and using jsonschema to validate the document.

Only the first error message is returned, which
might not be the best solution. When there are
multiple errors, the client must keep sending
the document and fixing the errors one by one.

A better structure would be to have a nested
structure with all the fields, and messages,
similar to what jsonschema returns, so the
client has a chance to fix all problems at
once.