Refactoring duplicated code
Moving type definitions to the new app
package, instead of having double
declaration in schema and validator
packages.

Checking if a document is valid JSON
got also moved to the app package.

Schema Id is validated by the schema
package (if it's empty), the implementation
in the validator package got deleted.