Logging a few errors
Log if a schema couldn't be saved to the
filesystem, or a file couldn't be read
for some reason (including the user asking
for a not-existing schema - probably this
case should be filtered out).

Invalid JSONs are also logged, which is
probably an overkill, but helps to find cases
when the validation works incorrectly.
This log should be disabled by default.

If the user's document is invalid according
to the schema, that's normal behavior of the
application, hence logging it is unnecessary.